Move TweakValue from Tweak.swift to TweakValue.swift
